Had int/float backwards in inspect page

+++ src/main/java/water/api/Inspect.java
@@ -205,7 +205,7 @@ public class Inspect extends Request {
       json.addProperty(MEAN, c._mean);
       json.addProperty(VARIANCE, c._sigma);
       json.addProperty(NUM_MISSING_VALUES, va._numrows - c._n);
-      json.addProperty(TYPE, c._domain != null ? "enum" : (c.isFloat() ? "int" : "float"));
+      json.addProperty(TYPE, c._domain != null ? "enum" : (c.isFloat() ? "float" : "int"));
       json.addProperty(ENUM_DOMAIN_SIZE, c._domain != null ? c._domain.length : 0);
       cols.add(json);
     }
